backend/requirements.txt
Django==3.2.25
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.10.1
django-filter==21.1
Pillow==8.4.0
PyJWT==1.7.1
gunicorn==20.1.0
whitenoise==6.0.0
python-dotenv==0.19.2
